pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "srt-model-quantizing"
version = "1.0.0"
description = "Enterprise-grade model optimization and serving platform"
authors = ["SolidRusT Networks <info@soln.ai>"]
license = "MIT"
readme = "README.md"
packages = [{include = "srt_model_quantizing"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
pydantic = "^1.10.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
prometheus-client = "^0.16.0"
opentelemetry-api = "^1.17.0"
opentelemetry-sdk = "^1.17.0"
opentelemetry-instrumentation-fastapi = "^0.38b0"
opentelemetry-exporter-jaeger = "^1.17.0"
structlog = "^23.1.0"
mlflow = "^2.3.1"
torch = "^2.0.0"
nemo-toolkit = {extras = ["all"], version = "^1.18.0"}
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.3.5"
pynvml = "^11.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.2.0"
httpx = "^0.24.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
pytest-env = "^0.8.1"
types-python-jose = "^3.3.4"
types-passlib = "^1.7.7"

[tool.poetry.scripts]
srt-serve = "srt_model_quantizing.cli:app serve"
srt-quantize = "srt_model_quantizing.cli:app"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "nemo.*",
    "mlflow.*",
    "opentelemetry.*",
    "prometheus_client.*",
    "pynvml.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=srt_model_quantizing --cov-report=term-missing"
testpaths = ["tests"]
env = [
    "ENV=test",
    "SECRET_KEY=test_secret_key",
    "MODEL_REPOSITORY=/tmp/test_models",
    "CATALOG_PATH=/tmp/test_catalog"
]
